CC-34135 Fixes after CR
