+++ composer.json
@@ -8,7 +8,6 @@
         "php": ">=8.1",
         "spryker/kernel": "^3.30.0",
         "spryker/symfony": "^3.0.0",
-        "spryker/translator": "^1.1.0",
         "spryker/user": "^3.0.0",
         "spryker/user-extension": "^1.4.0",
         "spryker/uuid-behavior": "^1.0.0"
