+++ data/translation/Zed/en_US.csv
