+++ phpstan.neon
@@ -1,6 +1,5 @@
 parameters:
     level: 8
-    checkGenericClassInNonGenericObjectType: false
     paths:
         - src/
     excludePaths:
