+++ src/SprykerEco/Zed/AmazonQuicksight/Communication/Expander/QuicksightUserFormExpander.php
@@ -8,7 +8,7 @@
 namespace SprykerEco\Zed\AmazonQuicksight\Communication\Expander;
 
 use SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig;
-use SprykerEco\Zed\AmazonQuicksight\Dependency\Facade\AmazonQuicksightToTranslatorFacadeInterface;
+use SprykerEco\Zed\AmazonQuicksight\Communication\DataProvider\QuicksightUserFormDataProviderInterface;
 use Symfony\Component\Form\DataTransformerInterface;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\FormBuilderInterface;
@@ -30,7 +30,7 @@ class QuicksightUserFormExpander implements QuicksightUserFormExpanderInterface
     /**
      * @var string
      */
-    protected const FIELD_QUICKSIGHT_USER_ROLE_HELP_MESSAGE = 'Author: A user who can create data sources, datasets, analyses, and dashboards. <br/>Reader: A user who has read-only access to dashboards.';
+    protected const HELP_MESSAGE_QUICKSIGHT_USER_ROLE = 'Author: A user who can create data sources, datasets, analyses, and dashboards. <br/>Reader: A user who has read-only access to dashboards.';
 
     /**
      * @var string
@@ -48,9 +48,14 @@ class QuicksightUserFormExpander implements QuicksightUserFormExpanderInterface
     protected const KEY_ROLE = 'role';
 
     /**
-     * @var \Symfony\Component\Form\DataTransformerInterface
+     * @var \Symfony\Component\Form\DataTransformerInterface<array<string, mixed>, array<string, mixed>>
      */
     protected DataTransformerInterface $quicksightUserRoleDataTransformer;
+
+    /**
+     * @var \SprykerEco\Zed\AmazonQuicksight\Communication\DataProvider\QuicksightUserFormDataProviderInterface
+     */
+    protected QuicksightUserFormDataProviderInterface $quicksightUserFormDataProvider;
 
     /**
      * @var \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig
@@ -58,23 +63,18 @@ class QuicksightUserFormExpander implements QuicksightUserFormExpanderInterface
     protected AmazonQuicksightConfig $amazonQuicksightConfig;
 
     /**
-     * @var \SprykerEco\Zed\AmazonQuicksight\Dependency\Facade\AmazonQuicksightToTranslatorFacadeInterface
-     */
-    protected AmazonQuicksightToTranslatorFacadeInterface $translatorFacade;
-
-    /**
-     * @param \Symfony\Component\Form\DataTransformerInterface $quicksightUserRoleDataTransformer
+     * @param \Symfony\Component\Form\DataTransformerInterface<array<string, mixed>, array<string, mixed>> $quicksightUserRoleDataTransformer
+     * @param \SprykerEco\Zed\AmazonQuicksight\Communication\DataProvider\QuicksightUserFormDataProviderInterface $quicksightUserFormDataProvider
      * @param \SprykerEco\Zed\AmazonQuicksight\AmazonQuicksightConfig $amazonQuicksightConfig
-     * @param \SprykerEco\Zed\AmazonQuicksight\Dependency\Facade\AmazonQuicksightToTranslatorFacadeInterface $translatorFacade
      */
     public function __construct(
         DataTransformerInterface $quicksightUserRoleDataTransformer,
-        AmazonQuicksightConfig $amazonQuicksightConfig,
-        AmazonQuicksightToTranslatorFacadeInterface $translatorFacade
+        QuicksightUserFormDataProviderInterface $quicksightUserFormDataProvider,
+        AmazonQuicksightConfig $amazonQuicksightConfig
     ) {
         $this->quicksightUserRoleDataTransformer = $quicksightUserRoleDataTransformer;
+        $this->quicksightUserFormDataProvider = $quicksightUserFormDataProvider;
         $this->amazonQuicksightConfig = $amazonQuicksightConfig;
-        $this->translatorFacade = $translatorFacade;
     }
 
     /**
@@ -90,41 +90,16 @@ class QuicksightUserFormExpander implements QuicksightUserFormExpanderInterface
 
             $event->getForm()->add(static::FIELD_QUICKSIGHT_USER_ROLE, ChoiceType::class, [
                 'required' => false,
-                'choices' => $this->getQuicksightUserRoleChoices(),
+                'choices' => $this->quicksightUserFormDataProvider->getQuicksightUserRoleChoices(),
                 'placeholder' => static::PLACEHOLDER_QUICKSIGHT_USER_ROLE,
                 'attr' => [
                     'template_path' => static::TEMPLATE_PATH_QUICKSIGHT_USER_ROLE,
                     'disabled' => $quicksightUserRole && !$this->amazonQuicksightConfig->isQuicksightUserRoleUpdateEnabled(),
                 ],
-                'help' => $this->translatorFacade->trans(static::FIELD_QUICKSIGHT_USER_ROLE_HELP_MESSAGE),
+                'help' => static::HELP_MESSAGE_QUICKSIGHT_USER_ROLE,
             ]);
         });
 
         $builder->addModelTransformer($this->quicksightUserRoleDataTransformer);
-    }
-
-    /**
-     * @return array<string, string>
-     */
-    protected function getQuicksightUserRoleChoices(): array
-    {
-        $quicksightUserRoles = $this->amazonQuicksightConfig->getQuicksightUserRoles();
-
-        return array_combine(
-            $this->formatQuicksightUserRoleChoiceLabels($quicksightUserRoles),
-            $quicksightUserRoles,
-        );
-    }
-
-    /**
-     * @param list<string> $quicksightUserRoles
-     *
-     * @return list<string>
-     */
-    protected function formatQuicksightUserRoleChoiceLabels(array $quicksightUserRoles): array
-    {
-        return array_map(function ($quicksightUserRole) {
-            return ucfirst(strtolower($quicksightUserRole));
-        }, $quicksightUserRoles);
     }
 }
