--- src/SprykerEco/Zed/AmazonQuicksight/Dependency/Facade/AmazonQuicksightToTranslatorFacadeInterface.php
@@ -1,21 +1,0 @@
-<?php
-
-/**
- * MIT License
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace SprykerEco\Zed\AmazonQuicksight\Dependency\Facade;
-
-interface AmazonQuicksightToTranslatorFacadeInterface
-{
-    /**
-     * @param string $id
-     * @param array<string, mixed> $parameters
-     * @param string|null $domain
-     * @param string|null $locale
-     *
-     * @return string
-     */
-    public function trans(string $id, array $parameters = [], ?string $domain = null, ?string $locale = null): string;
-}
